MSL: fix buffer ordering when transpiling compute from SPIRV (#77)
Fixes the buffer ordering when transpiling compute shaders
from SPIRV.
The SPV kernel execution model produces a different
layout for HLSL and GLSL. When used with GLSL, the bindings
set are ignored by SPIRV-cross. In this case, use GLCompute
model instead. This produces the expected result.
Also properly assigns the return values that are checked
for errors.

Signed-off-by: Mirko Covizzi <mrkcvzz@gmail.com>