Require Vulkan 1.1 for wave intrinsics
Shader Model 6.0 has support for GPU features such as wave intrinsics,
e.g. `WaveGetLaneCount`, `WaveActiveSum` and `WaveGetLaneIndex`.
Shadercross currently compiles HLSL shaders under SM 6.0.

In order to support cross-compilation of HLSL shaders utilising these
features to SPIR-V, spirv-cross requires a minimum of Vulkan 1.1. If
this is not set explicitly, it will complain and refuse to compile the
shader - this makes wave intrinsics unusable.

In order to specify the target environment, we need to pass
`-fspv-target-env` as a command line parameter to spirv-cross. This
commit passes Vulkan 1.1 as the target environment when compiling to
SPIR-V, to support these Shader Model 6.0 features.

I have tested this support with b0nes164's implementation of GPU Prefix
Sum[1], which utilizes wave intrinsics to increase performance, and with
Vulkan 1.1 required, these shaders compile to SPIR-V and execute
successfully.

  [1] https://github.com/b0nes164/GPUPrefixSums/tree/98d93a4e9ed2f3c8353119515bf9be90a2e137ad/GPUPrefixSumsD3D12